projeto-ms902-bot/requirements.txt
feedparser==6.0.2
beautifulsoup4==4.9.3
nltk==3.5
python-dateutil==2.8.1
mysql-connector-python==8.0.25